+++ src/Helpers/ConverterHelper.php
@@ -13,10 +13,10 @@ class ConverterHelper implements IConverter
 
     /**
      * @param string $value
-     * @return string string
+     * @return string
      * @ref https://stackoverflow.com/a/1993772/155905
      */
-    public function fromCamelCaseToSnakeCase(string $value)
+    public function fromCamelCaseToSnakeCase(string $value): string
     {
         if (empty($value)) {
             return '';
@@ -29,19 +29,19 @@ class ConverterHelper implements IConverter
         );
         $ret = $matches[0];
         foreach ($ret as &$match) {
-            $match = $match == strtoupper($match) ? strtolower($match) : lcfirst($match);
+            $match = $match === strtoupper($match) ? strtolower($match) : lcfirst($match);
         }
         return implode('_', $ret);
     }
 
     /**
-     * @param $object
+     * @param mixed $object
      * @param boolean $excludeEmptyProps
      * @return array
      */
-    public function fromObjectToArray($object, $excludeEmptyProps = false)
+    public function fromObjectToArray($object, $excludeEmptyProps = false): array
     {
-        if (empty($object) || !is_object($object)) {
+        if (empty($object) || !\is_object($object)) {
             return [];
         }
 
@@ -49,13 +49,13 @@ class ConverterHelper implements IConverter
     }
 
     /**
-     * @param $object
+     * @param mixed $object
      * @param boolean $excludeEmptyProps
      * @return string
      */
-    public function fromObjectToJson($object, $excludeEmptyProps = false)
+    public function fromObjectToJson($object, $excludeEmptyProps = false): string
     {
-        if (empty($object) || !is_object($object)) {
+        if (empty($object) || !\is_object($object)) {
             return '';
         }
 
@@ -63,12 +63,12 @@ class ConverterHelper implements IConverter
     }
 
     /**
-     * @param $object
+     * @param mixed $object
      * @param boolean $excludeEmptyProps
      * @return array
      * @ref https://stackoverflow.com/a/37610076/155905
      */
-    private function extractPropertiesFromObject($object, $excludeEmptyProps = false)
+    private function extractPropertiesFromObject($object, $excludeEmptyProps = false): array
     {
         $result = [];
 
@@ -81,7 +81,7 @@ class ConverterHelper implements IConverter
             }
             $refProperty = $reflector->getProperty($property->getName());
             $refProperty->setAccessible(true);
-            $result[self::fromCamelCaseToSnakeCase($property->getName())] = $property->getValue($object);
+            $result[$this->fromCamelCaseToSnakeCase($property->getName())] = $property->getValue($object);
         }
 
         return $result;