+++ src/IConverter.php
@@ -6,7 +6,7 @@ interface IConverter
 {
     public function fromCamelCaseToSnakeCase(string $value);
 
-    public function fromObjectToArray($object);
+    public function fromObjectToArray($object, $excludeEmptyProps = false);
 
-    public function fromObjectToJson($object);
+    public function fromObjectToJson($object, $excludeEmptyProps = false);
 }