+++ tests/Mocks/UserModelMock.php
@@ -1,0 +1,107 @@
+<?php
+
+namespace Lfbn\BaseModel\Tests\Mocks;
+
+use Lfbn\BaseModel\AbstractBaseModel;
+
+class UserModelMock extends AbstractBaseModel
+{
+
+    /**
+     * @var int
+     */
+    protected $id;
+
+    /**
+     * @var string
+     */
+    protected $name;
+
+    /**
+     * @var float
+     */
+    protected $height;
+
+    /**
+     * @var boolean
+     */
+    protected $active;
+
+    /**
+     * @return int
+     */
+    public function getId(): int
+    {
+        return $this->id;
+    }
+
+    /**
+     * @param int $id
+     */
+    public function setId(int $id): void
+    {
+        $this->id = $id;
+    }
+
+    /**
+     * @return string
+     */
+    public function getName(): string
+    {
+        return $this->name;
+    }
+
+    /**
+     * @param string $name
+     */
+    public function setName(string $name): void
+    {
+        $this->name = $name;
+    }
+
+    /**
+     * @return float
+     */
+    public function getHeight(): float
+    {
+        return $this->height;
+    }
+
+    /**
+     * @param float $height
+     */
+    public function setHeight(float $height): void
+    {
+        $this->height = $height;
+    }
+
+    /**
+     * @return bool
+     */
+    public function isActive(): bool
+    {
+        return $this->active;
+    }
+
+    /**
+     * @param bool $active
+     */
+    public function setActive(bool $active): void
+    {
+        $this->active = $active;
+    }
+
+    /**
+     * @return array
+     */
+    public function getValidationRules(): array
+    {
+        return [
+            ['property' => 'id', 'validator' => 'isNotEmpty'],
+            ['property' => 'id', 'validator' => 'isInteger'],
+            ['property' => 'name', 'validator' => 'isString'],
+            ['property' => 'height', 'validator' => 'isFloat'],
+            ['property' => 'active', 'validator' => 'isBoolean']
+        ];
+    }
+}